make config.py the script for running the config_to_json process
